# refactor-results/before-openapi.json
{
  "components": {
    "schemas": {
...
      "BeneficiaryDetail": {
        "additionalProperties": true,
        "properties": {
          "applications": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Applications",
            "type": "array"
          },
          "bank_account_id": {
            "title": "Bank Account Id",
            "type": "string"
          },
          "bank_account_id_node_id": {
            "title": "Bank Account Id Node Id",
            "type": "string"
          },
          "beneficiary_id": {
            "title": "Beneficiary Id",
            "type": "string"
          },
          "dob": {
            "title": "Dob",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "risk_level": {
            "enum": [
              "Low",
              "Medium",
              "High",
              "Critical"
            ],
            "title": "Risk Level",
            "type": "string"
          },
          "risk_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "type": "integer"
          }
        },
        "required": [
          "beneficiary_id",
          "full_name",
          "dob",
          "phone",
          "bank_account_id",
          "bank_account_id_node_id",
          "risk_score",
          "risk_level",
          "evidence"
        ],
        "title": "BeneficiaryDetail",
        "type": "object"
      },
...
      "Evidence": {
        "properties": {
          "beneficiary_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Id"
          },
          "contribution": {
            "title": "Contribution",
            "type": "integer"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "indicator": {
            "title": "Indicator",
            "type": "string"
          },
          "related_account_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Account Ids",
            "type": "array"
          },
          "related_beneficiary_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Beneficiary Ids",
            "type": "array"
          },
          "source_record_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Record Ids",
            "type": "array"
          }
        },
        "required": [
          "indicator",
          "contribution",
          "explanation",
          "related_beneficiary_ids",
          "related_account_ids",
          "source_record_ids"
        ],
        "title": "Evidence",
        "type": "object"
      },
      "FinancialFinding": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "assessment": {
            "enum": [
              "Review",
              "Context only"
            ],
            "title": "Assessment",
            "type": "string"
          },
          "beneficiary_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Id"
          },
          "beneficiary_ids": {
            "items": {
              "type": "string"
            },
            "title": "Beneficiary Ids",
            "type": "array"
          },
          "case_ids": {
            "items": {
              "type": "string"
            },
            "title": "Case Ids",
            "type": "array"
          },
          "contribution": {
            "title": "Contribution",
            "type": "integer"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "indicator": {
            "title": "Indicator",
            "type": "string"
          },
          "related_account_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Account Ids",
            "type": "array"
          },
          "related_beneficiary_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Beneficiary Ids",
            "type": "array"
          },
          "source_record_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Record Ids",
            "type": "array"
          },
          "transactions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "indicator",
          "contribution",
          "explanation",
          "related_beneficiary_ids",
          "related_account_ids",
          "source_record_ids",
          "finding_id",
          "beneficiary_ids",
          "case_ids",
          "transactions",
          "amount",
          "assessment"
        ],
        "title": "FinancialFinding",
        "type": "object"
      },
      "FindingPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FinancialFinding"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "FindingPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "Page",
        "type": "object"
      },
...
      "ReviewBeneficiary": {
        "additionalProperties": true,
        "properties": {
          "applications": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Applications",
            "type": "array"
          },
          "bank_account_id": {
            "title": "Bank Account Id",
            "type": "string"
          },
          "bank_account_id_node_id": {
            "title": "Bank Account Id Node Id",
            "type": "string"
          },
          "beneficiary_id": {
            "title": "Beneficiary Id",
            "type": "string"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "district": {
            "title": "District",
            "type": "string"
          },
          "dob": {
            "title": "Dob",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "institution_id": {
            "title": "Institution Id",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "risk_level": {
            "enum": [
              "Low",
              "Medium",
              "High",
              "Critical"
            ],
            "title": "Risk Level",
            "type": "string"
          },
          "risk_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "type": "integer"
          },
          "schemes": {
            "items": {
              "type": "string"
            },
            "title": "Schemes",
            "type": "array"
          },
          "total_disbursed": {
            "default": 0,
            "title": "Total Disbursed",
            "type": "number"
          }
        },
        "required": [
          "beneficiary_id",
          "full_name",
          "dob",
          "phone",
          "bank_account_id",
          "bank_account_id_node_id",
          "risk_score",
          "risk_level",
          "evidence",
          "district",
          "institution_id"
        ],
        "title": "ReviewBeneficiary",
        "type": "object"
      },
      "ReviewPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewBeneficiary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "ReviewPage",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local fictional scholarship audit prototype. Scores are review priorities, not fraud probabilities.",
    "title": "GrantLens",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/anomalies": {
      "get": {
        "operationId": "anomalies_api_v1_anomalies_get",
        "parameters": [
          {
            "in": "query",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 100,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "",
              "pattern": "^(|circular_transfer|collector|payout_concentration|overpayment)$",
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Anomalies"
      }
    },
    "/api/v1/anomalies/{finding_id}": {
      "get": {
        "operationId": "anomaly_api_v1_anomalies__finding_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialFinding"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Anomaly"
      }
    },
...
    "/api/v1/beneficiaries": {
      "get": {
        "operationId": "beneficiaries_api_v1_beneficiaries_get",
        "parameters": [
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "risk_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Beneficiaries"
      }
    },
    "/api/v1/beneficiaries/{beneficiary_id}": {
      "get": {
        "operationId": "beneficiary_api_v1_beneficiaries__beneficiary_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "beneficiary_id",
            "required": true,
            "schema": {
              "title": "Beneficiary Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeneficiaryDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Beneficiary"
      }
    },
    "/api/v1/beneficiaries/{beneficiary_id}/matches": {
      "get": {
        "operationId": "matches_api_v1_beneficiaries__beneficiary_id__matches_get",
        "parameters": [
          {
            "in": "path",
            "name": "beneficiary_id",
            "required": true,
            "schema": {
              "title": "Beneficiary Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Audit Id"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Matches"
      }
    },
    "/api/v1/beneficiaries/{beneficiary_id}/transactions": {
      "get": {
        "operationId": "beneficiary_transactions_api_v1_beneficiaries__beneficiary_id__transactions_get",
        "parameters": [
          {
            "in": "path",
            "name": "beneficiary_id",
            "required": true,
            "schema": {
              "title": "Beneficiary Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Beneficiary Transactions"
      }
    },
...
    "/api/v1/review-queue": {
      "get": {
        "operationId": "review_queue_api_v1_review_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 100,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "risk_level",
            "required": false,
            "schema": {
              "default": "",
              "pattern": "^(|Low|Medium|High|Critical)$",
              "title": "Risk Level",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "district",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 100,
              "title": "District",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "risk",
              "pattern": "^(risk|amount|name|id)$",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scheme",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 50,
              "title": "Scheme",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cluster_id",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 80,
              "title": "Cluster Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "case_status",
            "required": false,
            "schema": {
              "default": "",
              "pattern": "^(|Needs Review|In Investigation|Verification Requested|Cleared)$",
              "title": "Case Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Queue"
      }
    },
...
  }
}